2nd reg fix, changed code for cancel invite only if not expired #4460

+++ app/Controller/TeamsController.php
@@ -925,6 +925,15 @@ class TeamsController extends AppController
         $error_msg = '';
         $invite_data = $this->Team->Invite->findById($invite_id);
 
+        // if already expired throw error, you can't cancel
+        if ($action_flg == 'Canceled' && ($invite_data['Invite']['email_token_expires'] < REQUEST_TIMESTAMP)) {
+            $error_msg = (__("Error, this invitation already expired, you can't cancel."));
+            $res['title'] = $error_msg;
+            $res['error'] = true;
+            $this->Pnotify->outError($error_msg);
+            return $this->_ajaxGetResponse($res);
+        }
+
         // for cancel the old invite
         $res = $this->Team->Invite->delete($invite_id);
 
@@ -939,6 +948,7 @@ class TeamsController extends AppController
             if (!$invite) {
                 $error = true;
                 $error_msg = (__("Error, failed to invite."));
+                $this->Pnotify->outError($error_msg);
             }
             //send invite mail
             $team_name = $this->Team->TeamMember->myTeams[$this->Session->read('current_team_id')];
